requirements.txt
discord.py
discord.py[voice]
Pillow
PyNaCl
requests
aioredis==2.0.0b1
async-timeout
asyncio
discord-interactions